feat(apps/gql): move things around for better folder structure
