+++ apps/gql/src/loaders/index.ts
@@ -1,0 +1,1 @@
+export * from "./users";
