+++ apps/gql/src/schema/index.ts
@@ -1,0 +1,3 @@
+export * from "./types.generated";
+
+export type KampusResolvers = Resolvers<{ loaders: DataLoaders }>;
